backstabber/requirements.txt
Flask-Restless==0.17.0
Flask-SQLAlchemy==2.3.2
SQLAlchemy==1.2.5
requests==2.14.2
